Context: Make members final. Make event horizon default value 100 into static constant.
This change makes all fields in a Context instance final to simplify and make the rules clearer. The fields are not meant to be updated as the Context object is propagated in different threads as Thread Local Storage. Even if that mechanism gets removed later, different set of values (in the same process) should be represented as different Context instances.

(cherry picked from commit 814248a966798d1fa49ff2848ef10cf5f044a32c)