Remove the concept of "acceptable Base58 address headers" (version headers) for a given network.
The concept was always weakly defined, as version headers have been used
for different types ofdata too (which wasn't included in the list).
What's acceptable is always dependent on the usecase, not only the
network.

This is an API breaking change, but I doubt many users have used this
in the form of a list.

(cherry picked from commit ea2dc695d781a9f927f37f33810bf9b3713fd6ae)