DevNet:  Update DGW to match Dash Core behavior
Dash Core uses 256 bit integers for the difficulty math, which truncates the results.  DashJ uses BigIntegers which does not truncate.  After addition and multiplication operations, we need to check for larger than 256 bit numbers and remove the higher bytes.

This is due to a very high proof of work limit.

Signed-off-by: HashEngineering <hashengineeringsolutions@gmail.com>