Utils: Introduce Utils.resetMocking() and use it from AbstractFullPrunedBlockChainTest.
It appears that the tests do not run in the same order or the static variables
are not reset between runs on all systems or Gradle versions.
So it happened that the test UtilsTest.testRollMockClock() was messing up
the time for the H2FullPrunedBlockChainTest (and the LevelDB one), failing tests.